MA-212 Mobile API support for backward compatibility of course updates.
