+++ lms/djangoapps/mobile_api/course_info/views.py
@@ -7,6 +7,7 @@ from rest_framework.response import Response
 
 from courseware.courses import get_course_about_section, get_course_info_section_module
 from static_replace import make_static_urls_absolute, replace_static_urls
+from xmodule_modifiers import get_course_update_items
 
 from ..utils import MobileView, mobile_course_access
 
@@ -38,7 +39,7 @@ class CourseUpdatesList(generics.ListAPIView):
     @mobile_course_access()
     def list(self, request, course, *args, **kwargs):
         course_updates_module = get_course_info_section_module(request, course, 'updates')
-        update_items = reversed(getattr(course_updates_module, 'items', []))
+        update_items = list(reversed(get_course_update_items(course_updates_module)))
 
         updates_to_show = [
             update for update in update_items
